+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Support for Start/Stop `oci_integration_integration_instance`
+- Support compartmentId query for service marketplace
 
 ## 3.72.0 (April 22, 2020)